+++ activerecord/CHANGELOG.md
@@ -1,3 +1,9 @@
+*   Cache `CollectionAssociation#reader` proxy separately when a different
+    current_scope is active. This prevents the reader from retaining `.scoping`
+    information after the `.scoping` block has been exited.
+
+    *Ben Woosley*
+
 *   Increase pg gem version requirement to `~> 0.18`. Earlier versions of the
     pg gem are known to have problems with Ruby 2.2.